+++ core/app_services.go
@@ -132,10 +132,6 @@ func (app *application) serUpdateAccountPreferences(id string, appID string, org
 	return false, nil
 }
 
-func (app *application) serDeleteAccount(id string, apps []string) error {
-	return app.auth.DeleteAccount(id, apps)
-}
-
 func (app *application) serGetAccounts(limit int, offset int, appID string, orgID string, accountID *string, firstName *string, lastName *string, authType *string,
 	authTypeIdentifier *string, anonymous *bool, hasPermissions *bool, permissions []string, roleIDs []string, groupIDs []string) ([]model.Account, error) {
 	//find the accounts