+++ driven/storage/conversions_user.go
@@ -22,6 +22,7 @@ import (
 func orgAppMembershipFromStorage(item orgAppMembership, appOrg model.ApplicationOrganization) model.OrgAppMembership {
 	roles := accountRolesFromStorage(item.Roles, appOrg)
 	groups := accountGroupsFromStorage(item.Groups, appOrg)
+
 	return model.OrgAppMembership{ID: item.ID, AppOrg: appOrg, Permissions: item.Permissions,
 		Roles: roles, Groups: groups, Preferences: item.Preferences,
 		MostRecentClientVersion: item.MostRecentClientVersion}
@@ -58,10 +59,10 @@ func orgAppMembershipToStorage(item model.OrgAppMembership) orgAppMembership {
 	roles := accountRolesToStorage(item.Roles)
 	groups := accountGroupsToStorage(item.Groups)
 	preferences := item.Preferences
-	mostRecentClientVersions := item.MostRecentClientVersion
-	return orgAppMembership{ID: id, AppOrgID: appOrgID,
-		Permissions: permissions, Roles: roles, Groups: groups,
-		Preferences: preferences, MostRecentClientVersion: mostRecentClientVersions}
+	mostRecentClientVersion := item.MostRecentClientVersion
+
+	return orgAppMembership{ID: id, AppOrgID: appOrgID, Permissions: permissions, Roles: roles, Groups: groups,
+		Preferences: preferences, MostRecentClientVersion: mostRecentClientVersion}
 }
 
 func orgAppsMembershipsToStorage(items []model.OrgAppMembership) []orgAppMembership {
@@ -70,6 +71,33 @@ func orgAppsMembershipsToStorage(items []model.OrgAppMembership) []orgAppMembers
 		res[i] = orgAppMembershipToStorage(c)
 	}
 	return res
+}
+
+// DeletedOrgAppMembership
+func deletedMembershipsFromStorage(items []deletedOrgAppMembership, appsOrgs []model.ApplicationOrganization) []model.DeletedOrgAppMembership {
+	if len(items) == 0 {
+		return make([]model.DeletedOrgAppMembership, 0)
+	}
+
+	res := make([]model.DeletedOrgAppMembership, len(items))
+	for i, item := range items {
+		//find the application organization
+		appOrg := model.ApplicationOrganization{ID: item.AppOrgID}
+		for i := range appsOrgs {
+			cAppOrg := appsOrgs[i]
+			if cAppOrg.ID == item.AppOrgID {
+				appOrg = cAppOrg
+				break
+			}
+		}
+
+		res[i] = model.DeletedOrgAppMembership{ID: item.ID, AccountID: item.AccountID, AppOrg: appOrg, Context: item.Context, DateCreated: item.DateCreated}
+	}
+	return res
+}
+
+func deletedMembershipToStorage(item model.DeletedOrgAppMembership) deletedOrgAppMembership {
+	return deletedOrgAppMembership{ID: item.ID, AccountID: item.AccountID, AppOrgID: item.AppOrg.ID, Context: item.Context, DateCreated: item.DateCreated}
 }
 
 // Account