+++ driver/web/conversions_application.go
@@ -18,6 +18,9 @@ import (
 	"core-building-block/core/model"
 	Def "core-building-block/driver/web/docs/gen"
 	"core-building-block/utils"
+
+	"github.com/rokwire/logging-library-go/v2/errors"
+	"github.com/rokwire/logging-library-go/v2/logutils"
 )
 
 // PartialApp
@@ -30,6 +33,47 @@ func partialAppsToDef(item []model.Application) []Def.PartialApp {
 	result := make([]Def.PartialApp, len(item))
 	for i, item := range item {
 		result[i] = partialAppToDef(item)
+	}
+	return result
+}
+
+// DeletedMemberships
+func deletedMembershipsMapToDef(memberships map[model.AppOrgPair][]model.DeletedOrgAppMembership) []Def.DeletedMembership {
+	result := make([]Def.DeletedMembership, 0)
+	for appOrgPair, orgAppMemberships := range memberships {
+		result = append(result, Def.DeletedMembership{AppId: appOrgPair.AppID, OrgId: appOrgPair.OrgID, Memberships: deletedMembershipsToDef(orgAppMemberships)})
+	}
+	return result
+}
+
+// DeletedMembershipContext
+func deletedMembershipsFromDef(items []Def.DeletedMembershipContext) ([]model.DeletedOrgAppMembership, error) {
+	result := make([]model.DeletedOrgAppMembership, len(items))
+	for i, item := range items {
+		if item.AppId == nil {
+			return nil, errors.ErrorData(logutils.StatusMissing, "app_id", nil)
+		}
+		result[i] = deletedMembershipFromDef(item)
+	}
+	return result, nil
+}
+
+func deletedMembershipFromDef(item Def.DeletedMembershipContext) model.DeletedOrgAppMembership {
+	var context map[string]interface{}
+	if item.Context != nil {
+		context = *item.Context
+	}
+	return model.DeletedOrgAppMembership{AppOrg: model.ApplicationOrganization{Application: model.Application{ID: *item.AppId}}, Context: context}
+}
+
+func deletedMembershipsToDef(items []model.DeletedOrgAppMembership) []Def.DeletedMembershipContext {
+	result := make([]Def.DeletedMembershipContext, len(items))
+	for i, item := range items {
+		var context *map[string]interface{}
+		if item.Context != nil {
+			context = &item.Context
+		}
+		result[i] = Def.DeletedMembershipContext{AccountId: &item.AccountID, Context: context}
 	}
 	return result
 }