+++ driver/web/docs/gen/gen_types.go
@@ -331,6 +331,20 @@ type Config_Data struct {
 	union json.RawMessage
 }
 
+// DeletedMembership defines model for DeletedMembership.
+type DeletedMembership struct {
+	AppId       string                     `json:"app_id"`
+	Memberships []DeletedMembershipContext `json:"memberships"`
+	OrgId       string                     `json:"org_id"`
+}
+
+// DeletedMembershipContext defines model for DeletedMembershipContext.
+type DeletedMembershipContext struct {
+	AccountId *string                 `json:"account_id,omitempty"`
+	AppId     *string                 `json:"app_id,omitempty"`
+	Context   *map[string]interface{} `json:"context,omitempty"`
+}
+
 // Device defines model for Device.
 type Device struct {
 	DeviceId *string    `json:"device_id"`
@@ -1351,11 +1365,23 @@ type PostBbsAccountsCountParams struct {
 	OrgId *string `form:"org_id,omitempty" json:"org_id,omitempty"`
 }
 
+// GetBbsDeletedMembershipsParams defines parameters for GetBbsDeletedMemberships.
+type GetBbsDeletedMembershipsParams struct {
+	// ServiceId The ID of the service making the request
+	ServiceId string `form:"service_id" json:"service_id"`
+
+	// StartTime The start time for filtering results, specified as a Unix timestamp in seconds
+	StartTime *int64 `form:"start_time,omitempty" json:"start_time,omitempty"`
+}
+
 // GetBbsServiceRegsParams defines parameters for GetBbsServiceRegs.
 type GetBbsServiceRegsParams struct {
 	// Ids A comma-separated list of service IDs to return registrations for
 	Ids string `form:"ids" json:"ids"`
 }
+
+// DeleteServicesAccountJSONBody defines parameters for DeleteServicesAccount.
+type DeleteServicesAccountJSONBody = []DeletedMembershipContext
 
 // DeleteServicesAccountParams defines parameters for DeleteServicesAccount.
 type DeleteServicesAccountParams struct {
@@ -1727,6 +1753,9 @@ type PostBbsAccountsCountJSONRequestBody = PostBbsAccountsCountJSONBody
 
 // PostBbsServiceAccountIdJSONRequestBody defines body for PostBbsServiceAccountId for application/json ContentType.
 type PostBbsServiceAccountIdJSONRequestBody = ServicesReqServiceAccountsParams
+
+// DeleteServicesAccountJSONRequestBody defines body for DeleteServicesAccount for application/json ContentType.
+type DeleteServicesAccountJSONRequestBody = DeleteServicesAccountJSONBody
 
 // PostServicesAccountFollowJSONRequestBody defines body for PostServicesAccountFollow for application/json ContentType.
 type PostServicesAccountFollowJSONRequestBody = Follow